refactor(tag): defineOptions instead of two script once more
