+++ packages/nutui/components/tag/tag.ts
@@ -1,7 +1,7 @@
 import type { ExtractPropTypes } from 'vue'
 import { commonProps, makeNumericProp, makeStringProp } from '../_utils'
 import { CLICK_EVENT, CLOSE_EVENT } from '../_constants'
-import type { TagType } from './type'
+import type { TagType } from './types'
 
 export const tagProps = {
   ...commonProps,
